+++ src/main/java/roomescape/dto/ReservationReqDto.java
@@ -5,6 +5,6 @@ import java.time.LocalDate;
 public record ReservationReqDto(
         String name,
         LocalDate date,
-        int timeId
+        long timeId
 ) {
 }
